+++ ambra_sdk/__init__.py
@@ -1,11 +1,11 @@
-__version__ = '3.21.3.0-2'
+__version__ = '3.21.4.0-1'
 
 # from footer in api.html
 API_VERSION = 'LBL0022 v45.0 2021-04-07'
 # equal for API_VERSION based on dictionary.html
 MODELS_VERSION = 'LBL0022 v45.0 2021-04-07'
 # from storage_api.html
-STORAGE_VERSION = 'LBL0038 v14.0 2021-02-24'
+STORAGE_VERSION = 'LBL0038 v15.0 2021-05-19'
 
-DOCS_URL = 'https://dicomgrid.github.io/sdk-python/addon.html'
+DOCS_URL = 'https://dicomgrid.github.io/sdk-python'
 ADDON_DOCS_URL = '{docs_url}/addon.html'.format(docs_url=DOCS_URL)